Refactoring, creating separate TripService
