--- TourGuide/src/main/java/tourGuide/TourGuideModule.java
@@ -1,37 +1,0 @@
-package tourGuide;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import gpsUtil.GpsUtil;
-import rewardCentral.RewardCentral;
-import tourGuide.service.GpsService;
-import tourGuide.service.RewardsService;
-import tourGuide.service.UserService;
-import tourGuide.user.User;
-
-@Configuration
-public class TourGuideModule {
-	@Autowired
-	UserService userService;
-
-	@Autowired
-	GpsService gpsService;
-	
-	@Bean
-	public GpsUtil getGpsUtil() {
-		return new GpsUtil();
-	}
-	
-	@Bean
-	public RewardsService getRewardsService() {
-		return new RewardsService(gpsService, getRewardCentral(), userService);
-	}
-	
-	@Bean
-	public RewardCentral getRewardCentral() {
-		return new RewardCentral();
-	}
-	
-}
