+++ TourGuide/src/main/resources/application.properties
@@ -1,2 +1,3 @@
 
 logging.level.tourGuide=DEBUG
+tripPricer.api.key=test-server-api-key
