+++ TourGuide/src/test/java/tourGuide/TestPerformance.java
@@ -16,11 +16,9 @@ import gpsUtil.location.Attraction;
 import gpsUtil.location.VisitedLocation;
 import rewardCentral.RewardCentral;
 import tourGuide.helper.InternalTestHelper;
-import tourGuide.service.GpsService;
-import tourGuide.service.RewardsService;
-import tourGuide.service.TourGuideService;
-import tourGuide.service.UserService;
+import tourGuide.service.*;
 import tourGuide.user.User;
+import tripPricer.TripPricer;
 
 
 public class TestPerformance {
@@ -49,13 +47,13 @@ public class TestPerformance {
 	
 	@Test
 	public void highVolumeTrackLocationConc() {
-		GpsUtil gpsUtil = new GpsUtil();
-		GpsService gpsService = new GpsService(gpsUtil);
+		GpsService gpsService = new GpsService(new GpsUtil());
 		UserService userService = new UserService();
+		TripService tripService = new TripService(new TripPricer());
 		RewardsService rewardsService = new RewardsService(gpsService, new RewardCentral(), userService);
 		// Users should be incremented up to 100,000, and test finishes within 15 minutes
 		InternalTestHelper.setInternalUserNumber(NUMBER_OF_TEST_USERS);
-		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService);
+		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService, tripService);
 
 		StopWatch stopWatch = new StopWatch();
 		stopWatch.start();
@@ -69,13 +67,13 @@ public class TestPerformance {
 
 	@Test
 	public void highVolumeTrackLocationAndProcessConc() {
-		GpsUtil gpsUtil = new GpsUtil();
-		GpsService gpsService = new GpsService(gpsUtil);
+		GpsService gpsService = new GpsService(new GpsUtil());
 		UserService userService = new UserService();
+		TripService tripService = new TripService(new TripPricer());
 		RewardsService rewardsService = new RewardsService(gpsService, new RewardCentral(), userService);
 		// Users should be incremented up to 100,000, and test finishes within 15 minutes
 		InternalTestHelper.setInternalUserNumber(NUMBER_OF_TEST_USERS);
-		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService);
+		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService, tripService);
 
 		StopWatch stopWatch = new StopWatch();
 		stopWatch.start();
@@ -92,13 +90,14 @@ public class TestPerformance {
 	public void highVolumeGetRewards() {
 		GpsService gpsService = new GpsService(new GpsUtil());
 		UserService userService = new UserService();
+		TripService tripService = new TripService(new TripPricer());
 		RewardsService rewardsService = new RewardsService(gpsService, new RewardCentral(), userService);
 
 		// Users should be incremented up to 100,000, and test finishes within 20 minutes
 		InternalTestHelper.setInternalUserNumber(NUMBER_OF_TEST_USERS);
 		StopWatch stopWatch = new StopWatch();
 		stopWatch.start();
-		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService);
+		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService, tripService);
 
 	    Attraction attraction = gpsService.getAttractions().get(0);
 		List<User> allUsers = new ArrayList<>();
@@ -123,13 +122,14 @@ public class TestPerformance {
 	public void highVolumeGetRewardsCallConc() {
 		GpsService gpsService = new GpsService(new GpsUtil());
 		UserService userService = new UserService();
+		TripService tripService = new TripService(new TripPricer());
 		RewardsService rewardsService = new RewardsService(gpsService, new RewardCentral(), userService);
 
 		// Users should be incremented up to 100,000, and test finishes within 20 minutes
 		InternalTestHelper.setInternalUserNumber(NUMBER_OF_TEST_USERS);
 		StopWatch stopWatch = new StopWatch();
 		stopWatch.start();
-		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService);
+		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService, tripService);
 
 		Attraction attraction = gpsService.getAttractions().get(0);
 		List<User> allUsers = new ArrayList<>();
@@ -180,13 +180,14 @@ public class TestPerformance {
 	public void highVolumeGetRewardsOneCall() {
 		GpsService gpsService = new GpsService(new GpsUtil());
 		UserService userService = new UserService();
+		TripService tripService = new TripService(new TripPricer());
 		RewardsService rewardsService = new RewardsService(gpsService, new RewardCentral(), userService);
 
 		// Users should be incremented up to 100,000, and test finishes within 20 minutes
 		InternalTestHelper.setInternalUserNumber(NUMBER_OF_TEST_USERS);
 		StopWatch stopWatch = new StopWatch();
 		stopWatch.start();
-		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService);
+		TourGuideService tourGuideService = new TourGuideService(gpsService, rewardsService, userService, tripService);
 
 		System.out.println("Starting Adding Locations");
